Refactor. Reorder the execution order of the rules...
At this point in exercise, we should break if more than 1 source file exists (its unexpected), if the file is in lowercase (non following java conventions), if there is a java file with any other name than Rectangle.java...

Finally if we do find Rectangle.java, we're good.

In any other scenario, like a typo, Rectngle.java, we break saying it could possibly be because of unknown reasons... From resource bundle we should say to recheck if the trainee feels this is what is expected of the requirement and if yes then check with mentor.